Database setup

+++ components/FavouriteButton/FavouriteButton.js
@@ -6,49 +6,65 @@ export default function FavouriteButton( {
   image,
   idDrink,
   id,
-  mutate
+  mutate,
+  favourites
 } ) {
   const {data: session, status} = useSession();
   const router = useRouter();
+  const favouriteItem = favourites && Array.isArray(favourites) ? favourites.find(fav => fav.idDrink === idDrink) : null;
 
   async function handleToggleFavorite() {
-
-    if (session) {
-      const favourite = {
-        idDrink: idDrink,
-        strDrink: name,
-        strDrinkThumb: image,
-        userId: session.user.id
-      }
-        console.log(favourite)
-        const response = await fetch("/api/favourites", {
-            method: "POST",
-            headers: {
-              "Content-Type": "application/json",
-            },
-            body: JSON.stringify(favourite),
-          });
+    if (!session) {
+      signIn();
+      return;
     }
-    else {
-            signIn();
-        }
+
+    const favourite = {
+      idDrink: idDrink,
+      strDrink: name,
+      strDrinkThumb: image,
+      userId: session.user.id
     };
 
-    async function handleDelete() {
-      const response = await fetch(`/api/favourites/${id}`, {
-        method: "DELETE",
-    })
-    if (response.ok){
+    console.log(favourite);
+    const response = await fetch("/api/favourites", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(favourite),
+    });
+
+    if (response.status === 409) {
+      alert('This drink is already in your favorites!');
+    } else if (response.ok) {
       mutate();
+    } else {
+      alert('There was an error processing your request.');
     }
-    };
+  }
 
-      return (
-        <div>
-        <button className="favButton" onClick={handleToggleFavorite}>🤍
-        {/* <span>{foundCocktailInfo?.isFavorite ? "❤️" : "🤍"}</span> */}
-        </button>
-        <button onClick={handleDelete}>Delete</button>
-        </div>
-      )
+  async function handleDelete() {
+    if (!favouriteItem) return;
+
+    const response = await fetch(`/api/favourites/${favouriteItem._id}`, {
+      method: "DELETE",
+    });
+
+    if (response.ok) {
+      mutate();
+    } else {
+      console.error("Failed to delete the favourite");
+    }
+  }
+
+  return (
+    <div>
+      {favouriteItem ? (
+        <button onClick={handleDelete}>❤️</button>
+      ) : (
+        <button className="favButton" onClick={handleToggleFavorite}>🤍</button>
+      )}
+    </div>
+  );
 }
